pyproject.toml
[project]
name = "times-doctor"
version = "0.6.4"
description = "Diagnose and scan failed TIMES/Veda runs; optional LLM advice"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name = "David Green", email = "david.green@csiro.au"}
]
keywords = ["TIMES", "VEDA", "energy-modeling", "GAMS", "CPLEX", "optimization"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering",
  "Typing :: Typed",
]

dependencies = [
  "typer",
  "rich",
  "python-dotenv",
  "httpx",
  "psutil",
  "tomli; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-cov>=4.0",
  "pytest-mock>=3.10",
  "pytest-env>=0.8",
  "respx>=0.21.0",
  "ruff>=0.8.0",
  "mypy>=1.0",
  "pre-commit>=3.0",
  "detect-secrets>=1.5.0",
]

[project.scripts]
times-doctor = "times_doctor.cli:app"

[project.urls]
Homepage = "https://github.com/dlg0/times-doctor"
Repository = "https://github.com/dlg0/times-doctor"
Issues = "https://github.com/dlg0/times-doctor/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/times_doctor"]

[[tool.uv.index]]
url = "https://pypi.org/simple"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "S", "PERF"]
ignore = [
    "E501",  # Line too long (enforced by formatter, not linter)
    "E722",  # Bare except (acceptable in robust error handling)
    "S110",  # try-except-pass (acceptable for optional cleanup)
    "S112",  # try-except-continue (acceptable for lenient parsing)
    "B904",  # raise without from (acceptable for user-facing errors)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "SIM117"]  # Allow assert and nested with in tests
"scripts/*" = ["B007"]  # Allow unused loop vars in scripts
"src/times_doctor/core/http_client.py" = ["S311"]  # random.uniform for retry jitter (not crypto)
"src/times_doctor/core/llm.py" = ["S603", "PERF401", "SIM102"]  # subprocess for llm cli, allow simple loops
"src/times_doctor/core/qa_check_parser.py" = ["PERF401", "PERF403"]  # Comprehensions harder to read here
"src/times_doctor/core/lst_parser.py" = ["PERF401"]  # List comprehensions harder to read here
"src/times_doctor/cli.py" = ["B008", "S603", "PERF203", "E741"]  # typer patterns, subprocess trusted, perf ok

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
disallow_incomplete_defs = true
no_implicit_optional = true
exclude = ["tests/", "docs/", ".venv/"]

[dependency-groups]
dev = [
    "detect-secrets>=1.5.0",
]
